Initial Commit
Got a basic two person chatter working. Need to implement encryption all the way through, then some way to sneak the key over to the second user. I'll be displaying the message as the server see's it. The goal is to make messages only legible by the client for full, guaranteed, privacy